// src/app/model/user-activity.ts
export interface UserActivity {
  notificationId?: string;
  notification?: string;
  insuredTp?: string;
  companyLogo?: string;
  notificationDate?: Date;
  activity?: Date;
  policyCarId?: string;
  notificationNature?: string;
  notificationStatus?: string;
  notificationStatusCode?: string;
  expertName?: string;
  carsCount?: string;
  injuryCount?: string;
  accidentLocation?: string;
  policyType?: string;
  brandTrademark?: string;
  plate?: string;
  yom?: string;
  ownerName?: string;
  insCompany?: string;
  userName?: string;

  notes?: string;
  type?: string;
}